Major change to code that waits for ESP response
Changed from serial.find to waitForString function, that actively
listens for the target string, or exists after timeout.
Now it should take around 5 seconds to send the message after wake.
Moved a lot of string to PROGMEN memory.
Change speed of debug console to be as fast a possible.
Running the clean function in just one place now.